Walking_the_Approximate_Longest_Path
Jenna is playing a computer game involving a large map with n cities numbered sequentially from  1 to n  that are connected by m bidirectional roads. The game's objective is to travel to as many cities as possible without visiting any city more than once. The more cities the player visits, the more points they earn.
Take input of number of cities, number of ways and list of all ways and print the length of longest path along with the order of the cities visited.